requirements.txt
replicate
langchain 
python-dotenv
google-generativeai
langchain-google-genai
PyPDF2
streamlit
openai
tiktoken
unstructured
pdf2image
pdfminer
playsound
python-dotenv
langchain-google-genai
PyPDF2
gTTS
flask
googletrans
python-docx
streamlit-carousel
faiss-cpu
numpy
Pillow
easyocr
torchvision
torch
firebase-admin
streamlit_mic_recorder
streamlit-chat 